Ported the path for "DeleteAsset()". Adds a new option to allow remote asset deletion in robust handler.
Ported r/12732, except that the method remained  DeleteAsset(UUID)
rather than Delete(string), with all references changed accordingly.

+++ OpenSim/Services/AssetService/AssetService.cs
@@ -157,7 +157,14 @@ namespace OpenSim.Services.AssetService
             if (!UUID.TryParse(id, out assetID))
                 return false;
 
-            return m_Database.DeleteAsset(assetID);
+            AssetMetadata meta = m_Database.GetMetadata(assetID);
+            if (meta == null)
+                return false;
+
+            if ((int)(meta.Flags & AssetFlags.Maptile) != 0)
+                return m_Database.DeleteAsset(assetID);
+
+            return false;
         }
 
         void HandleShowDigest(string module, string[] args)
